% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasaurus_dozen}
\alias{datasaurus_dozen}
\title{datasaurus_dozen data}
\description{
datasaurus dozen data from datasauRus pacakge,
all plots have same x and y mean, variance and correlation,
but look different visually. variables for dataset, x and y
}
