% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datasaurus_dozen_wide}
\alias{datasaurus_dozen_wide}
\title{datasaurus_dozen_wide data}
\description{
datasaurus dozen data from datasauRus pacakge, but instead
of having a variable for which dataset, each dataset has its own
x and y variables
}
