% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{features}
\alias{features}
\title{Simulated data with special features}
\format{
A tibble with 1,013 observations and 3 variables,
and 15 different patterns
\describe{
\item{feature}{label of data set}
\item{x}{variable for horizontal axis}
\item{y}{variable for vertical axis}
}
}
\description{
Simulated data with common features that might
be seen in 2D data. Variable are feature, x, y.
}
