% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustscags.R
\name{sc_clumpy_robust}
\alias{sc_clumpy_robust}
\title{Compute robust clumpy scagnostic measure using MST}
\usage{
sc_clumpy_robust(x, y)
}
\description{
Compute robust clumpy scagnostic measure using MST
}
\examples{
  require(ggplot2)
  require(tidyr)
  require(dplyr)
  ggplot(datasaurus_dozen, aes(x=x, y=y)) +
    geom_point() +
    facet_wrap(~dataset, ncol=3, scales = "free")
  sc_clumpy_robust(datasaurus_dozen_wide$away_x, datasaurus_dozen_wide$away_y)
  sc_clumpy_robust(datasaurus_dozen_wide$dots_x, datasaurus_dozen_wide$dots_y)
}
