% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alphahull.R
\name{sc_convex}
\alias{sc_convex}
\alias{sc_convex.scree}
\alias{sc_convex.default}
\alias{sc_convex.list}
\title{Compute convex scagnostic measures}
\usage{
sc_convex(x, y)

\method{sc_convex}{scree}(x, y = NULL)

\method{sc_convex}{default}(x, y)

\method{sc_convex}{list}(chull, ahull)
}
\description{
Compute convex scagnostic measures
}
\examples{
  require(ggplot2)
  require(tidyr)
  require(dplyr)
  data(anscombe)
  anscombe_tidy <- anscombe \%>\%
  pivot_longer(cols = everything(),
    names_to = c(".value", "set"),
    names_pattern = "(.)(.)")
  ggplot(anscombe_tidy, aes(x=x, y=y)) +
    geom_point() +
    facet_wrap(~set, ncol=2, scales = "free")
  a_s <- scree(anscombe$x1, anscombe$y1)
  sc_convex(a_s)
  a_s <- scree(anscombe$x2, anscombe$y2)
  sc_convex(a_s)
  a_s <- scree(anscombe$x3, anscombe$y3)
  sc_convex(a_s)
  a_s <- scree(anscombe$x4, anscombe$y4)
  sc_convex(a_s)
}
