% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{sc_dcor}
\alias{sc_dcor}
\title{Distance correlation index.}
\usage{
sc_dcor(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
A "numeric" object that gives the plot's dcor score.
}
\description{
The distance correlation between X and Y defined by Székely, et al.
in Measuring and testing dependence by correlation of distances.
The measure was suggested as an  association scagnostic in
Katrin Grimm's PhD thesis (2016). Distance correlation is a measure
of non-linear dependence which is 0 if and only if the two variables
are independent. It is computed using an ANOVA like calculation on the
pairwise distances between observations.
}
\examples{
  require(ggplot2)
  require(tidyr)
  require(dplyr)
  data(anscombe)
  anscombe_tidy <- anscombe |>
  pivot_longer(cols = everything(),
    names_to = c(".value", "set"),
    names_pattern = "(.)(.)")
  ggplot(anscombe_tidy, aes(x=x, y=y)) +
    geom_point() +
    facet_wrap(~set, ncol=2, scales = "free")
  sc_dcor(anscombe$x1, anscombe$y1)
  sc_dcor(anscombe$x2, anscombe$y2)
  sc_dcor(anscombe$x3, anscombe$y3)
  sc_dcor(anscombe$x4, anscombe$y4)

}
