% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted.R
\name{sc_grid}
\alias{sc_grid}
\alias{sc_grid.default}
\alias{sc_grid.scree}
\alias{sc_grid.igraph}
\title{Compute the grid scanostic measure using MST}
\usage{
sc_grid(x, y, epsilon)

\method{sc_grid}{default}(x, y, epsilon = 0.01)

\method{sc_grid}{scree}(x, y = NULL, epsilon = 0.01)

\method{sc_grid}{igraph}(x, y, epsilon = 0.01)
}
\arguments{
\item{x}{numeric vector of x values, or an MST object}

\item{y}{numeric vector of y values, or a scree object}

\item{epsilon}{the error tolerance allowed to count the angles}
}
\value{
The plot's grid score as a numeric object
}
\description{
This scagnostic identifies grid-like structures by counting the
number of 90 and 180 degree angles in the MST, as defined in
in Adam Rahman's PhD thesis.
}
\examples{
  require(ggplot2)
  require(dplyr)
  ggplot(features, aes(x=x, y=y)) +
     geom_point() +
     facet_wrap(~feature, ncol = 5, scales = "free")
  features |> group_by(feature) |>
    summarise(grid1 = sc_grid(x,y),
              grid2 = sc_grid(x,y, epsilon=0.05))
  sc_grid(datasaurus_dozen_wide$away_x, datasaurus_dozen_wide$away_y)

}
