% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{sc_outlying}
\alias{sc_outlying}
\alias{sc_outlying.default}
\alias{sc_outlying.scree}
\title{Compute outlying scagnostic measure using MST}
\usage{
sc_outlying(x, y)

\method{sc_outlying}{default}(x, y)

\method{sc_outlying}{scree}(sc)
}
\arguments{
\item{x}{numeric vector of x values}

\item{y}{numeric vector of y values}

\item{sc}{scree object if pre-built}

\item{mst}{mst object if pre-built}
}
\description{
Compute outlying scagnostic measure using MST
}
\examples{
  require(ggplot2)
  require(tidyr)
  require(dplyr)
  ggplot(datasaurus_dozen, aes(x=x, y=y)) +
    geom_point() +
    facet_wrap(~dataset, ncol=3, scales = "free")
  sc_outlying(datasaurus_dozen_wide$dino_x, datasaurus_dozen_wide$dino_y)
  sc_outlying(datasaurus_dozen_wide$dots_x, datasaurus_dozen_wide$dots_y)
  sc_outlying(datasaurus_dozen_wide$h_lines_x, datasaurus_dozen_wide$h_lines_y)

}
