% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryfunctions.R
\name{sc_pairwise}
\alias{sc_pairwise}
\title{Compute scagnostics on all possible scatter plots for the given data}
\usage{
sc_pairwise(
  all_data,
  scags = c("outlying", "stringy", "striated", "clumpy", "sparse", "skewed", "convex",
    "skinny", "monotonic", "splines", "dcor"),
  euclid = TRUE
)
}
\description{
Compute scagnostics on all possible scatter plots for the given data
}
\examples{
require(GGally)
require(dplyr)
# Calculate selected scagnostics
sc_pairwise(datasaurus_dozen_wide,
  scags=c("outlying","clumpy","monotonic"))

# Calculate all scagnsotics for all variable pairs and plot it
p <- sc_pairwise(datasaurus_dozen_wide) \%>\%
  GGally::ggpairs(scagdata, columns = 3:ncol(scagdata))
p
# to make it interactive use
# plotly::ggplotly(p)

}
\seealso{
calc_scags
}
