% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{sc_sparse}
\alias{sc_sparse}
\title{Compute sparse scagnostic measure using MST}
\usage{
sc_sparse(x, y, out.rm = TRUE, binner = "hex")
}
\arguments{
\item{x, y}{numeric vectors}

\item{out.rm}{logical; if TRUE, iteratively trim large MST edges, If FALSE
the scanostics will be computed on the entire data set with no outlier
removal.}

\item{binner}{an optional function that bins the x and y vectors prior
to triangulation
Can be:
\itemize{
\item \code{"hex"} (default): hexagonal binning following the procedure in the
graph-theoretic scagnostics paper (start 40x40, halve
until <= 250 nonempty cells)
\item \code{NULL}: no binning (use raw points)
\item a function: user-defined binner
}}
}
\value{
A numeric object that gives the plot's sparse score.
}
\description{
Identifies if the data is confined to a small number of locations on the
plane. It was first defined in Scagnostics Distributions by
Wilkinson & Wills (2008). It is calculated as the 90th percentile
of MST edge lengths
}
\examples{
  require(ggplot2)
  require(tidyr)
  require(dplyr)
  ggplot(datasaurus_dozen, aes(x=x, y=y)) +
    geom_point() +
    facet_wrap(~dataset, ncol=3, scales = "free")
  sc_sparse(datasaurus_dozen_wide$away_x, datasaurus_dozen_wide$away_y)
  sc_sparse(datasaurus_dozen_wide$circle_x, datasaurus_dozen_wide$circle_y)
  sc_sparse(datasaurus_dozen_wide$dino_x, datasaurus_dozen_wide$dino_y)

}
