% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted.R
\name{sc_sparse2}
\alias{sc_sparse2}
\title{Compute  adjusted sparse measure using the alpha hull}
\usage{
sc_sparse2(x, y, out.rm = TRUE, binner = "hex")
}
\arguments{
\item{x, y}{numeric vectors}

\item{out.rm}{logical; if TRUE, iteratively trim large MST edges, If FALSE
the scanostics will be computed on the entire data set with no outlier
removal.}

\item{binner}{an optional function that bins the x and y vectors prior
to triangulation
Can be:
\itemize{
\item \code{"hex"} (default): hexagonal binning following the procedure in the
graph-theoretic scagnostics paper (start 40x40, halve
until <= 250 nonempty cells)
\item \code{NULL}: no binning (use raw points)
\item a function: user-defined binner
}}
}
\value{
A numeric object that gives the plot's adjusted sparse score.
}
\description{
The sparse2 mesure as defined in cassowaryr, Mason, et al (2025).
The measure calculates the sparsity of the plot as 1-area(ahull).
}
\examples{
  require(ggplot2)
  require(tidyr)
  require(dplyr)
  data(anscombe_tidy)
  ggplot(anscombe_tidy, aes(x=x, y=y)) +
    geom_point() +
    facet_wrap(~set, ncol=2, scales = "free")
  sc_sparse2(anscombe$x1, anscombe$y1)

}
