% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/association.R
\name{sc_splines}
\alias{sc_splines}
\title{Spline based index.}
\usage{
sc_splines(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
A "numeric" object that gives the plot's spines score.
}
\description{
Measures the functional non-linear dependence by fitting a penalised
splines model on X using Y, and on Y using X. The measure was
defined as an association scagnostic in Katrin Grimm's PhD thesis (2016)
The variance of the residuals are scaled down by the axis so they are
comparable, and finally the maximum is taken.
Therefore the value will be closer to 1 if either relationship
can be decently explained by a splines model.
}
\examples{
  require(ggplot2)
  require(tidyr)
  require(dplyr)
  data(anscombe)
  anscombe_tidy <- anscombe \%>\%
  pivot_longer(cols = everything(),
    names_to = c(".value", "set"),
    names_pattern = "(.)(.)")
  ggplot(anscombe_tidy, aes(x=x, y=y)) +
    geom_point() +
    facet_wrap(~set, ncol=2, scales = "free")
  sc_splines(anscombe$x1, anscombe$y1)
  sc_splines(anscombe$x2, anscombe$y2)
  sc_splines(anscombe$x3, anscombe$y3)

}
