% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{sc_striated_adjusted}
\alias{sc_striated_adjusted}
\alias{sc_striated_adjusted.scree}
\alias{sc_striated_adjusted.default}
\alias{sc_striated_adjusted.igraph}
\title{Compute angle adjusted stirated measure using MST}
\usage{
sc_striated_adjusted(x, y)

\method{sc_striated_adjusted}{scree}(x, y = NULL)

\method{sc_striated_adjusted}{default}(x, y)

\method{sc_striated_adjusted}{igraph}(mst, x)
}
\description{
Compute angle adjusted stirated measure using MST
}
\examples{
  require(ggplot2)
  require(tidyr)
  require(dplyr)
  data(anscombe_tidy)
  ggplot(anscombe_tidy, aes(x=x, y=y)) +
    geom_point() +
    facet_wrap(~set, ncol=2, scales = "free")
  sc_striated_adjusted(anscombe$x1, anscombe$y1)
}
