% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{sc_stringy}
\alias{sc_stringy}
\alias{sc_stringy.default}
\alias{sc_stringy.scree}
\alias{sc_stringy.igraph}
\alias{sc_stringy2}
\alias{sc_stringy2.default}
\alias{sc_stringy2.scree}
\alias{sc_stringy2.igraph}
\title{Compute stringy scagnostic measure using MST}
\usage{
sc_stringy(x, y, outlier_rm = FALSE, binner = NULL)

\method{sc_stringy}{default}(x, y, outlier_rm = FALSE, binner = NULL)

\method{sc_stringy}{scree}(x, y = NULL, outlier_rm = FALSE, binner = NULL)

\method{sc_stringy}{igraph}(x, y = NULL, outlier_rm = FALSE, binner = NULL)

sc_stringy2(x, y, outlier_rm = FALSE, binner = NULL)

\method{sc_stringy2}{default}(x, y, outlier_rm = FALSE, binner = NULL)

\method{sc_stringy2}{scree}(x, y = NULL, outlier_rm = FALSE, binner = NULL)

\method{sc_stringy2}{igraph}(x, y = NULL, outlier_rm = FALSE, binner = NULL)
}
\arguments{
\item{x}{numeric vector of x values}

\item{y}{numeric vector of y values}

\item{outlier_rm}{logical; if TRUE, iteratively trim large MST edges}

\item{binner}{an optional function that bins the x and y vectors prior
to triangulation
Can be:
\itemize{
\item \code{NULL}: no binning (use raw points)
\item \code{"hex"} : hexagonal binning following the procedure in the
graph-theoretic scagnostics paper (start 40x40, halve
until <= 250 nonempty cells)
\item a function: user-defined binner
}}
}
\value{
A "numeric" object that gives the plot's stringy score.
}
\description{
Compute stringy scagnostic measure using MST
}
\examples{
  require(ggplot2)
  require(tidyr)
  require(dplyr)
  data(anscombe_tidy)
  ggplot(anscombe_tidy, aes(x=x, y=y)) +
    geom_point() +
    facet_wrap(~set, ncol=2, scales = "free")
  sc_stringy(anscombe$x1, anscombe$y1)
  sc_stringy(anscombe$x2, anscombe$y2)
  sc_stringy(anscombe$x3, anscombe$y3)
  sc_stringy(anscombe$x4, anscombe$y4)

}
