% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{sc_stringy05}
\alias{sc_stringy05}
\title{Compute the stringy05 scagnostic measure}
\usage{
sc_stringy05(x, y, out.rm = TRUE, binner = "hex")
}
\arguments{
\item{x, y}{numeric vectors}

\item{out.rm}{logical; if TRUE, iteratively trim large MST edges, If FALSE
the scagnostics will be computed on the entire data set with no outlier
removal.}

\item{binner}{an optional function that bins the x and y vectors prior
to triangulation
Can be:
\itemize{
\item \code{"hex"} (default): hexagonal binning following the procedure in the
graph-theoretic scagnostics paper (start 40x40, halve
until <= 250 nonempty cells)
\item \code{NULL}: no binning (use raw points)
\item a function: user-defined binner
}}
}
\value{
A numeric value giving the stringy05 score.
}
\description{
Computes the stringy measure as defined in \emph{Graph-Theoretic Scagnostics}
(Wilkinson et al., 2005). It is the legnth of the longest shortest path through
the MST divided by the sum of all edge lengths in the MST.
}
\examples{
x <- datasaurus_dozen_wide$star_x
y <- datasaurus_dozen_wide$star_y
sc_stringy05(x, y)

}
