% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{sc_stringy06}
\alias{sc_stringy06}
\title{Compute the stringy06 scagnostic measure using the MST}
\usage{
sc_stringy06(x, y, out.rm = TRUE, binner = "hex")
}
\arguments{
\item{x, y}{numeric vectors}

\item{out.rm}{logical; if TRUE, iteratively trim large MST edges, If FALSE
the scagnostics will be computed on the entire data set with no outlier
removal.}

\item{binner}{an optional function that bins the x and y vectors prior
to triangulation
Can be:
\itemize{
\item \code{"hex"} (default): hexagonal binning following the procedure in the
graph-theoretic scagnostics paper (start 40x40, halve
until <= 250 nonempty cells)
\item \code{NULL}: no binning (use raw points)
\item a function: user-defined binner
}}
}
\value{
A numeric object that gives the plot's stringy score.
}
\description{
This measure identifies a “stringy” shape with no branches, such as a thin
line of data. The stringy06 function is defined in
\emph{High-Dimensional Visual Analytics: Interactive Exploration Guided by Pairwise
Views of Point Distributions} (Wilkinson et al., 2006). It is calculated using
the minimum spanning tree (MST) by comparing the number of vertices with
degree two to the total number of vertices, dropping those of degree one.
}
\details{
The name "stringy06" is used to distinguish this version from earlier definition
of the stringy measure.
}
\examples{
require(ggplot2)
require(dplyr)

# plot the features data
ggplot(features, aes(x=x, y=y)) +
  geom_point() +
  facet_wrap(~feature, ncol = 5, scales = "free")

# calculate using tidy code
features |>
 group_by(feature) |>
 summarise(stringy = sc_stringy06(x,y))

# using just vectors of points
x <- datasaurus_dozen_wide$star_x
y <- datasaurus_dozen_wide$star_y

# plot it
ggplot() +
  geom_point(aes(x = x, y = y))

# calculate using vectors
sc_stringy06(x, y)
}
