% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{sc_striped}
\alias{sc_striped}
\title{Measure of Discreteness}
\usage{
sc_striped(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
double
}
\description{
This metric computes the 1-(ratio between the
number of unique values to total data values)
on number of rotations of the data, and
returns the smallest value. If this value is
large it means that there are only a few unique
data values, and hence the distribution is
discrete
}
\examples{
data("datasaurus_dozen_wide")
sc_striped(datasaurus_dozen_wide$v_lines_x,
           datasaurus_dozen_wide$v_lines_y)
sc_striped(datasaurus_dozen_wide$dino_x,
           datasaurus_dozen_wide$dino_y)
}
