% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{top_functions}
\alias{top_functions}
\alias{top_pair}
\alias{top_scag}
\title{Summary computations for scagnostic data}
\usage{
top_pair(scags_data)

top_scag(scags_data)
}
\arguments{
\item{scags_data}{A dataset of scagnostic values that was returned by
calc_scags or calc_scags_wide}
}
\value{
A data frame. For top_pair, each row will represent a scagnostic
with its highest pair. For top_scag, each row will represent a pair of
variables with its highest valued scagnostic.
}
\description{
These functions suggests a summary statistic that can be found using
the scag calculations provided by calc_scags. The top_pair function
finds the top pair of variables for each of the scagnostics, while
top_scag finds the highest value scagnostic for each pair of variables.
While these computations
are relatively straight forward for any R user to compute themselves,
including these summary function in the package simultaneously streamlines
a common calculation made with the scagnostic data and suggests this
summary to new users of the package.
}
\examples{
require(dplyr)
# calculate scag data
scag_data <- datasaurus_dozen |>
  group_by(dataset) |>
  summarise(calc_scags(x,y, scags=c("monotonic", "outlying", "convex")))

# Calculate top_pair
scag_data |>
  top_pair()

# Calculate top_scag
scag_data |>
  top_scag()


}
\seealso{
calc_scags calc_scags_wide
}
