% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calc_scags_wide}
\alias{calc_scags_wide}
\title{Compute scagnostics on all possible scatter plots for the given data}
\usage{
calc_scags_wide(
  all_data,
  scags = c("outlying", "stringy", "striated", "striated_adjusted", "clumpy", "sparse",
    "skewed", "convex", "skinny", "monotonic", "splines", "dcor"),
  euclid = TRUE
)
}
\description{
Compute scagnostics on all possible scatter plots for the given data
}
\examples{
# Calculate selected scagnostics
calc_scags_wide(pk[,2:5], scags=c("outlying","clumpy","monotonic"))

}
\seealso{
calc_scags
}
