% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_convexhull}
\alias{draw_convexhull}
\title{Drawing the Convex Hull}
\usage{
draw_convexhull(x, y, alpha = 0.5, clr = "black", fill = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{clr}{optional colour of points and lines, default black}

\item{fill}{Fill the polygon}
}
\description{
This function will draw the Convex Hull for a
scatterplot.
}
\examples{
require(dplyr)
require(ggplot2)
data("features")
nl <- features \%>\% filter(feature == "nonlinear2")
draw_convexhull(nl$x, nl$y)
}
