% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_MST}
\alias{draw_MST}
\title{Drawing the MST}
\usage{
draw_MST(x, y, alpha = 0.5)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\description{
This function will draw the MST for a
scatterplot.
}
\examples{
require(dplyr)
require(ggplot2)
data("features")
nl <- features \%>\% filter(feature == "nonlinear2")
draw_MST(nl$x, nl$y)
}
