% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw_mst}
\alias{draw_mst}
\title{Drawing the MST}
\usage{
draw_mst(x, y, alpha = 0.5, out.rm = TRUE)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}

\item{alpha}{The alpha value used to build the graph object. Larger values allow points further apart to be connected.}

\item{out.rm}{option to return the outlier removed MST}
}
\value{
A "gg" object that draws the plot's MST.
}
\description{
This function will draw the MST for a
scatterplot.
}
\examples{
require(dplyr)
require(ggplot2)
data("features")
nl <- features \%>\% filter(feature == "nonlinear2")
draw_mst(nl$x, nl$y)
}
