% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted.R
\name{sc_clumpy2}
\alias{sc_clumpy2}
\title{Compute adjusted clumpy measure using MST}
\usage{
sc_clumpy2(x, y, out.rm = TRUE, binner = "hex")
}
\arguments{
\item{x, y}{numeric vectors}

\item{out.rm}{logical; if TRUE, iteratively trim large MST edges, If FALSE
the scanostics will be computed on the entire data set with no outlier
removal.}

\item{binner}{an optional function that bins the x and y vectors prior
to triangulation
Can be:
\itemize{
\item \code{"hex"} (default): hexagonal binning following the procedure in the
graph-theoretic scagnostics paper (start 40x40, halve
until <= 250 nonempty cells)
\item \code{NULL}: no binning (use raw points)
\item a function: user-defined binner
}}
}
\value{
A numeric object that gives the plot's adjusted clumpy score.
}
\description{
This measure is defined in the cassowaryr paper by Mason, et al. (2025).
It is an alternative measure for clumpiness. It is the ratio of the
between cluster edges and the within cluster edges. It is a good
alternative measure to clumpy when binning is removed as a pre-processing
step.
}
\examples{
  require(ggplot2)
  require(dplyr)
  ggplot(features, aes(x=x, y=y)) +
     geom_point() +
     facet_wrap(~feature, ncol = 5, scales = "free")
  features |> group_by(feature) |> summarise(clumpy = sc_clumpy2(x,y))
  sc_clumpy2(datasaurus_dozen_wide$away_x, datasaurus_dozen_wide$away_y)

}
