% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted.R
\name{sc_clumpy2}
\alias{sc_clumpy2}
\alias{sc_clumpy2.default}
\alias{sc_clumpy2.scree}
\alias{sc_clumpy2.igraph}
\title{Compute angle adjusted clumpy measure using MST}
\usage{
sc_clumpy2(x, y)

\method{sc_clumpy2}{default}(x, y)

\method{sc_clumpy2}{scree}(x, y = NULL)

\method{sc_clumpy2}{igraph}(mst, sc)
}
\arguments{
\item{x}{numeric vector of x values}

\item{y}{numeric vector of y values}
}
\description{
Compute angle adjusted clumpy measure using MST
}
\examples{
  require(ggplot2)
  require(tidyr)
  require(dplyr)
  data(anscombe_tidy)
  ggplot(anscombe_tidy, aes(x=x, y=y)) +
    geom_point() +
    facet_wrap(~set, ncol=2, scales = "free")
  sc_clumpy2(anscombe$x1, anscombe$y1)

}
