% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustscags.R
\name{sc_clumpy_r}
\alias{sc_clumpy_r}
\title{Compute robust clumpy scagnostic measure using MST}
\usage{
sc_clumpy_r(x, y, out.rm = TRUE, binner = "hex")
}
\arguments{
\item{x, y}{numeric vectors}

\item{out.rm}{logical; if TRUE, iteratively trim large MST edges, If FALSE
the scanostics will be computed on the entire data set with no outlier
removal.}

\item{binner}{an optional function that bins the x and y vectors prior
to triangulation
Can be:
\itemize{
\item \code{"hex"} (default): hexagonal binning following the procedure in the
graph-theoretic scagnostics paper (start 40x40, halve
until <= 250 nonempty cells)
\item \code{NULL}: no binning (use raw points)
\item a function: user-defined binner
}}
}
\value{
A numeric object that gives the plot's robust clumpy score.
}
\description{
A computation for clumpy that is supposed to make the measure more robust
to changes in binning. The scagnostic is defined in Improving the
Robustness of Scagnostics, Wang, et al. (2020).
}
\examples{
  require(ggplot2)
  require(dplyr)
  ggplot(features, aes(x=x, y=y)) +
     geom_point() +
     facet_wrap(~feature, ncol = 5, scales = "free")
  features \%>\% group_by(feature) \%>\% summarise(clumpy = sc_clumpy_r(x,y))
  sc_clumpy_r(datasaurus_dozen_wide$away_x, datasaurus_dozen_wide$away_y)

}
