% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted.R
\name{sc_grid}
\alias{sc_grid}
\title{Compute the grid scanostic measure using MST}
\usage{
sc_grid(x, y, epsilon, out.rm = TRUE, binner = "hex")
}
\arguments{
\item{x, y}{numeric vectors}

\item{epsilon}{the error tolerance allowed when deciding if the MST angles
are at a right angle or not}

\item{out.rm}{logical; if TRUE, iteratively trim large MST edges, If FALSE
the scanostics will be computed on the entire data set with no outlier
removal.}

\item{binner}{an optional function that bins the x and y vectors prior
to triangulation
Can be:
\itemize{
\item \code{"hex"} (default): hexagonal binning following the procedure in the
graph-theoretic scagnostics paper (start 40x40, halve
until <= 250 nonempty cells)
\item \code{NULL}: no binning (use raw points)
\item a function: user-defined binner
}}
}
\value{
A numeric object that gives the plot's grid score.
}
\description{
The grid scagnsotic as defined in Adam Rahman's PhD thesis (2018).
The scagnostic identifies grid-like structures by counting the number of
90 and 180 degree angles in the MST. This measure can be used as an
effective alternative to striated when computing scagnostics without binning.
}
\examples{
  require(ggplot2)
  require(dplyr)
  ggplot(features, aes(x=x, y=y)) +
     geom_point() +
     facet_wrap(~feature, ncol = 5, scales = "free")
  features |> group_by(feature) |>
    summarise(grid1 = sc_grid(x,y),
              grid2 = sc_grid(x,y, epsilon=0.05))
  sc_grid(datasaurus_dozen_wide$away_x, datasaurus_dozen_wide$away_y)

}
