% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{sc_skewed}
\alias{sc_skewed}
\title{Compute skewed scagnostic measure using MST}
\usage{
sc_skewed(x, y, out.rm = TRUE, binner = "hex")
}
\arguments{
\item{x, y}{numeric vectors}

\item{out.rm}{logical; if TRUE, iteratively trim large MST edges, If FALSE
the scanostics will be computed on the entire data set with no outlier
removal.}

\item{binner}{an optional function that bins the x and y vectors prior
to triangulation
Can be:
\itemize{
\item \code{"hex"} (default): hexagonal binning following the procedure in the
graph-theoretic scagnostics paper (start 40x40, halve
until <= 250 nonempty cells)
\item \code{NULL}: no binning (use raw points)
\item a function: user-defined binner
}}
}
\value{
A numeric object that gives the plot's skewed score.
}
\description{
A measure of skewness in the edge lengths of the MST (not in the
distribution of the data). It was first defined in Graph Theoretic
Scagnostics, Wilkinson, et al. (2005). It is the ratio between
the 90th to 50th percentile range and the 10th to 90th percentile range.
}
\examples{
require(ggplot2)
require(dplyr)

# plot the feature
ggplot(features, aes(x=x, y=y)) +
   geom_point() +
   facet_wrap(~feature, ncol = 5, scales = "free")

# calculate using tidy code
features |>
  group_by(feature) |>
  summarise(skewed = sc_skewed(x,y))

# using two vectors
x <- datasaurus_dozen_wide$away_x
y <- datasaurus_dozen_wide$away_y

# plot it
ggplot() +
  geom_point(aes(x = x, y = y))

# calculate using vectors
sc_skewed(x, y)

}
