% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{sc_striated}
\alias{sc_striated}
\alias{sc_striated.default}
\alias{sc_striated.scree}
\alias{sc_striated.igraph}
\title{Compute striated scagnostic measure using MST}
\usage{
sc_striated(x, y, outlier_rm = FALSE, binner = NULL)

\method{sc_striated}{default}(x, y, outlier_rm = FALSE, binner = NULL)

\method{sc_striated}{scree}(x, y = NULL, outlier_rm = FALSE, binner = NULL)

\method{sc_striated}{igraph}(x, y, outlier_rm = FALSE, binner = NULL)
}
\arguments{
\item{x}{numeric vector of x values}

\item{y}{numeric vector of y values}

\item{outlier_rm}{logical; if TRUE, iteratively trim large MST edges}

\item{binner}{an optional function that bins the x and y vectors prior
to triangulation
Can be:
\itemize{
\item \code{NULL}: no binning (use raw points)
\item \code{"hex"} : hexagonal binning following the procedure in the
graph-theoretic scagnostics paper (start 40x40, halve
until <= 250 nonempty cells)
\item a function: user-defined binner
}}
}
\value{
A "numeric" object that gives the plot's striated score.
}
\description{
Compute striated scagnostic measure using MST
}
\examples{
  require(ggplot2)
  require(dplyr)
  data(anscombe_tidy)
  ggplot(anscombe_tidy, aes(x=x, y=y)) +
    geom_point() +
    facet_wrap(~set, ncol=2, scales = "free")
  sc_striated(anscombe$x1, anscombe$y1)
  sc_striated(anscombe$x2, anscombe$y2)

}
