% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{sc_striated}
\alias{sc_striated}
\alias{sc_striated.scree}
\alias{sc_striated.default}
\alias{sc_striated.igraph}
\title{Compute stirated scagnostic measure using MST}
\usage{
sc_striated(x, y)

\method{sc_striated}{scree}(x, y = NULL)

\method{sc_striated}{default}(x, y)

\method{sc_striated}{igraph}(mst, x)
}
\description{
Compute stirated scagnostic measure using MST
}
\examples{
  require(ggplot2)
  require(tidyr)
  require(dplyr)
  data(anscombe_tidy)
  ggplot(anscombe_tidy, aes(x=x, y=y)) +
    geom_point() +
    facet_wrap(~set, ncol=2, scales = "free")
  sc_striated(anscombe$x1, anscombe$y1)
  sc_striated(anscombe$x2, anscombe$y2)
  sc_striated(anscombe$x3, anscombe$y3)
  sc_striated(anscombe$x4, anscombe$y4)

  ggplot(datasaurus_dozen, aes(x=x, y=y)) +
    geom_point() +
    facet_wrap(~dataset, ncol=3, scales = "free")
  sc_striated(datasaurus_dozen_wide$away_x, datasaurus_dozen_wide$away_y)
  sc_striated(datasaurus_dozen_wide$dino_x, datasaurus_dozen_wide$dino_y)
  sc_striated(datasaurus_dozen_wide$dots_x, datasaurus_dozen_wide$dots_y)
  sc_striated(datasaurus_dozen_wide$h_lines_x, datasaurus_dozen_wide$h_lines_y)

}
