% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mst.R
\name{sc_striated}
\alias{sc_striated}
\alias{sc_striated.default}
\alias{sc_striated.scree}
\alias{sc_striated.list}
\title{Compute stirated scagnostic measure using MST}
\usage{
sc_striated(x, y)

\method{sc_striated}{default}(x, y)

\method{sc_striated}{scree}(sc)

\method{sc_striated}{list}(mst, sc)
}
\arguments{
\item{x}{numeric vector of x values}

\item{y}{numeric vector of y values}

\item{sc}{scree object if pre-built}

\item{mst}{mst object if pre-built}
}
\description{
Compute stirated scagnostic measure using MST
}
\examples{
  require(ggplot2)
  require(dplyr)
  ggplot(features, aes(x=x, y=y)) +
     geom_point() +
     facet_wrap(~feature, ncol = 5, scales = "free")
  features \%>\% group_by(feature) \%>\% summarise(striated = sc_striated(x,y))
  sc_striated(datasaurus_dozen_wide$away_x, datasaurus_dozen_wide$away_y)

}
