% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjusted.R
\name{sc_striated2}
\alias{sc_striated2}
\alias{sc_striated2.scree}
\alias{sc_striated2.default}
\alias{sc_striated2.igraph}
\title{Compute angle adjusted stirated measure using MST}
\usage{
sc_striated2(x, y)

\method{sc_striated2}{scree}(x, y = NULL)

\method{sc_striated2}{default}(x, y)

\method{sc_striated2}{igraph}(mst, x)
}
\arguments{
\item{x}{numeric vector of x values}

\item{y}{numeric vector of y values}
}
\description{
Compute angle adjusted stirated measure using MST
}
\examples{
  require(ggplot2)
  require(tidyr)
  require(dplyr)
  data(anscombe_tidy)
  ggplot(anscombe_tidy, aes(x=x, y=y)) +
    geom_point() +
    facet_wrap(~set, ncol=2, scales = "free")
  sc_striated2(anscombe$x1, anscombe$y1)
}
